Rename src/flow.js to src/types.js
