+++ src/index.js
@@ -12,10 +12,10 @@ import type {FusionPlugin, Token} from 'fusion-core';
 import serverPlugin from './node';
 import clientPlugin, {HydrationStateToken} from './browser';
 import createI18nLoader from './loader';
-import type {I18nDepsType, I18nServiceType} from './flow.js';
+import type {I18nDepsType, I18nServiceType} from './types.js';
 import {I18nLoaderToken} from './tokens.js';
 
-export type {I18nServiceType} from './flow.js';
+export type {I18nServiceType};
 const I18nToken: Token<I18nServiceType> = createToken('I18nToken');
 
 const plugin: FusionPlugin<I18nDepsType, I18nServiceType> = __NODE__
